Extract UV_Utils to tests/uv_utils.py and rename test file to test_uv_smoke.py
